Invalid module structure for keystore Moved methods & added tests
